Revert "Fix double to float max limit"
This reverts commit 227d65108b8ae35a5bdf7d7e00679b8cd7c4c76e.

+++ assignment_7/src/Mesh.cpp
@@ -204,8 +204,8 @@ void Mesh::compute_normals()
 
 void Mesh::compute_bounding_box()
 {
-    bb_min_ = vec3(std::numeric_limits<float>::max());
-    bb_max_ = vec3(std::numeric_limits<float>::lowest());
+    bb_min_ = vec3(std::numeric_limits<double>::max());
+    bb_max_ = vec3(std::numeric_limits<double>::lowest());
 
     for (Vertex v: vertices_)
     {
